[#7] refactor :recycle: ImageCacheManager 생성 및 역할 분리
