+++ iOS/SideDishApp/SideDishApp/ViewControllers/MenuViewController.swift
@@ -12,7 +12,7 @@ import Toaster
 class MenuViewController: UIViewController, UITableViewDataSource, UITableViewDelegate {
     
     static let networkManager = NetworkManager()
-    static let imageCache = NSCache<NSString, UIImage>()
+    static let imageCacheManager = ImageCacheManager()
     private let sections:[String] = ["국","찌개","반찬"]
     @IBOutlet var tableView: MenuTableView!
